+++ oscsend.h
@@ -36,28 +36,28 @@ class OscSend : public QObject,
   Q_OBJECT
 public:
   // cstr 1: 0 NOOP, 1 PLAY, 2 PAUSE, 3 REWIND, 4 QUIT
-  OscSend(champMM champ, int time = 0);
+  OscSend(champMM champ, double time = 0);
   // cstr 2: 5 P_NAME, 10 P_URI, 11 P_COLOR, 12 M_NAME, 22 R_P_OPACITY, 23 R_P_VOLUME,
   // 24 R_P_RATE, 28 R_M_OPACITY, 32 R_M_DEPTH
-  OscSend(champMM champ, int unInt, QString name, int time = 0);
+  OscSend(champMM champ, int unInt, QString name, double time = 0);
   // cstr 3: 6 P_REWIND
-  OscSend(champMM champ, int p_ID1, int time);
+  OscSend(champMM champ, int p_ID1, double time);
   // cstr 4: 7 P_OPACITY, 8 P_VOLUME, 9 P_RATE, 13 M_OPACITY, 17 M_DEPTH, 18 P_XFADE
-  OscSend(champMM champ, int ID1, int var, int time);
+  OscSend(champMM champ, int ID1, int var, double time);
   // cstr 5: 14 M_VISIBLE, 15 M_SOLO, 16 M_LOCK, 19 P_FADE
-  OscSend(champMM champ, int ID1, bool isproperty, int time = 0);
+  OscSend(champMM champ, int ID1, bool isproperty, double time = 0);
   // cstr 6 : 21 R_P_REWIND
-  OscSend(champMM champ, QString name, int time = 0);
+  OscSend(champMM champ, QString name, double time = 0);
   // cstr 7 : 20 R_P_NAME, 25 R_P_URI, 26 R_P_COLOR, 27 R_M_NAME, 34 R_P_XFADE
-  OscSend(champMM champ, QString name1, QString name2, int time = 0);
+  OscSend(champMM champ, QString name1, QString name2, double time = 0);
   // CSTR 8 : 29 R_M_VISIBLE, 30 R_M_SOLO, 31 R_M_LOCK, 33 R_P_FADE
-  OscSend(champMM champ, QString name, bool isproperty, int time = 0);
+  OscSend(champMM champ, QString name, bool isproperty, double time = 0);
 
   void ExecuteSend();
-  void ExecuteXFade(int ID1, int ID2, int time);
-  void ExecuteFade(int ID1, int time, bool isfadein);
-  void ExecutePXFade(QString p_name1, QString p_name2, int time);
-  void ExecutePFade(QString p_name, int time, bool isfadein);
+  void ExecuteXFade(int ID1, int ID2, double time);
+  void ExecuteFade(int ID1, double time, bool isfadein);
+  void ExecutePXFade(QString p_name1, QString p_name2, double time);
+  void ExecutePFade(QString p_name, double time, bool isfadein);
 
   champMM m_champ;
   // paint var
@@ -80,7 +80,7 @@ public:
   bool m_m_islocked = false;
   int m_m_depth = 0;
   // general var
-  int m_time = 0;
+  double m_time = 0;
   bool m_isfadein = false; // pour P_FADE
   bool m_iswaiting = true; // pour savoir si la cue doit s'enchaîner
 };
